simple/requirements.txt
python-pptx
pillow
pdf2image
